fix: rename & split utils package, replace `iota`
