+++ internal/cluster/datasource_hagroups.go
@@ -15,7 +15,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 
-	"github.com/bpg/terraform-provider-proxmox/internal/tffwk"
+	"github.com/bpg/terraform-provider-proxmox/internal/structure"
 	"github.com/bpg/terraform-provider-proxmox/proxmox"
 	hagroups "github.com/bpg/terraform-provider-proxmox/proxmox/cluster/ha/groups"
 )
@@ -56,7 +56,7 @@ func (d *hagroupsDatasource) Schema(_ context.Context, _ datasource.SchemaReques
 	resp.Schema = schema.Schema{
 		Description: "Retrieves the list of High Availability groups.",
 		Attributes: map[string]schema.Attribute{
-			"id": tffwk.IDAttribute(),
+			"id": structure.IDAttribute(),
 			"group_ids": schema.SetAttribute{
 				Description: "The identifiers of the High Availability groups.",
 				ElementType: types.StringType,
