+++ internal/cluster/resource_hagroup.go
@@ -23,7 +23,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 
-	"github.com/bpg/terraform-provider-proxmox/internal/tffwk"
+	"github.com/bpg/terraform-provider-proxmox/internal/structure"
 	"github.com/bpg/terraform-provider-proxmox/proxmox"
 	hagroups "github.com/bpg/terraform-provider-proxmox/proxmox/cluster/ha/groups"
 )
@@ -63,7 +63,7 @@ func (r *hagroupResource) Schema(
 	resp.Schema = schema.Schema{
 		Description: "Manages a High Availability group in a Proxmox VE cluster.",
 		Attributes: map[string]schema.Attribute{
-			"id": tffwk.IDAttribute(),
+			"id": structure.IDAttribute(),
 			"group": schema.StringAttribute{
 				Description: "The identifier of the High Availability group to manage.",
 				Required:    true,
